backend/requirements.txt
Django==5.2.7
djangorestframework==4.3.1
django-cors-headers==4.4.0
